Clean up kernel.c file
Main kernel entry file was getting too bloated from the
diagnostic and experimental code. Now diagnostic code has
been moved to the diag/ directory, and various headers have
been moved around and cleaned up (especially interrupt code
moved to arch/x86_64/interrupt.h).
