+++ src/kernel/diag/sys.h
@@ -1,0 +1,11 @@
+/**
+ * Diagnostics to print out system information.
+ */
+#ifndef DIAG_SYS_H
+#define DIAG_SYS_H
+
+void run_printf_tests(void);
+void print_gdtr_info(void);
+void print_idtr_info(void);
+
+#endif
